requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
docarray==0.40.0
duckduckgo-search==8.0.2
executing==2.1.0
fastapi==0.115.8
fonttools==4.58.1
frozenlist==1.5.0
google-search-results==2.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.2.6
lazy-loader==0.4
lxml==5.4.0
markdown-it-py==3.0.0
marshmallow==3.23.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.1
openai==1.58.1
orjson==3.10.12
ormsgpack==1.10.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
primp==0.15.0
prompt-toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.0
pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.36
stack-data==0.6.3
starlette==0.45.3
tenacity==9.0.0
tifffile==2025.5.26
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20241016
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
